instead of having user provide the inbound ip/hostname, just ask a remote server what it should be
if you are running psiturk on a home internet connection behind a router, your "internal" ip may be different than the external one.  thus, the server 'host' might be 192.168.1.8 but it is accessed on the internet via a different ip.  to avoid the confusion in all cases the inbound ip is just set here by what an external observer _thinks_ is you ip.  you are of course responsible for port-forwarding, etc...  may want to mention this in the docs